🐛 Fix: 수정 newsSeq 자료형

See merge request s06-bigdata-dist-sub2/S06P22A406!26

+++ backend/src/main/java/com/newsbig/sinmunmul/repository/ScrapRepositorySupport.java
@@ -39,7 +39,7 @@ public class ScrapRepositorySupport {
 		return new PageImpl<>(scraps, pageable, scraps.size());
 	}
 	
-	public boolean checkScrap(int userSeq, int newsSeq) {
+	public boolean checkScrap(int userSeq, long newsSeq) {
 		try {
 			jpaQueryFactory.select(qScrap)
 				.from(qScrap)
@@ -53,7 +53,7 @@ public class ScrapRepositorySupport {
 		return true;
 	}
 	
-	public void deleteScrap(int userSeq, int newsSeq) {
+	public void deleteScrap(int userSeq, long newsSeq) {
 		List<Scrap> scraps = jpaQueryFactory.select(qScrap)
 				.from(qScrap)
 					.leftJoin(qUser).on(qUser.userSeq.eq(qScrap.user.userSeq))
